requirements.txt
netifaces
dnspython
requests
webscreenshot
python-nmap
progress
pythonping
python-gnupg